Add new layout components Add components to the Layout component Move public images to the public directory Update the README file
