+++ src/assets/stylesheets/styles.scss
@@ -15,6 +15,8 @@ body {
   background-image: url("/assets/images/background.png");
 }
 
+/* Standard */
+
 /* Headings */
 h1 {
   font-family: $heading-text-font-family;
