+++ src/layout/Body/Body.tsx
@@ -1,0 +1,8 @@
+import {ReactElement} from "react";
+import {Outlet} from "react-router";
+
+export const Body = (): ReactElement => {
+    return (
+        <Outlet/>
+    )
+}
