+++ src/layout/Body/index.ts
@@ -1,0 +1,1 @@
+export {Body} from "./Body.tsx";
