+++ src/layout/Footer/Footer.tsx
@@ -1,0 +1,7 @@
+import {ReactElement} from "react";
+
+export const Footer = (): ReactElement => {
+    return (
+        <></>
+    )
+}
