+++ src/layout/Header/Header.tsx
@@ -1,0 +1,7 @@
+import {ReactElement} from "react";
+
+export const Header = (): ReactElement => {
+    return (
+        <></>
+    )
+}
