+++ src/layout/Layouts.module.scss
@@ -2,8 +2,3 @@
   width: 100%;
   height: 100%;
 }
-
-main {
-  width: 100%;
-  height: 100%;
-}
