+++ src/layout/Layouts.tsx
@@ -1,13 +1,22 @@
 import {ReactElement} from "react";
-import {Outlet} from "react-router";
+import {Header} from "./Header";
+import {Footer} from "./Footer";
+import {Body} from "./Body";
 import classes from "./Layouts.module.scss";
 
 export const Layout = (): ReactElement => {
     return (
         <div className={classes.layout}>
+            <header>
+                <Header/>
+            </header>
             <main>
-                <Outlet/>
+                <Body/>
             </main>
+            <footer>
+                <Footer/>
+            </footer>
         </div>
     )
 }
+
